Renamed TextInputMapping to ModalMapping
