+++ src/main/java/net/dv8tion/jda/api/events/interaction/ModalSubmitInteractionEvent.java
@@ -19,7 +19,7 @@ package net.dv8tion.jda.api.events.interaction;
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.interactions.InteractionHook;
 import net.dv8tion.jda.api.interactions.ModalInteraction;
-import net.dv8tion.jda.api.interactions.modals.TextInputMapping;
+import net.dv8tion.jda.api.interactions.modals.ModalMapping;
 import net.dv8tion.jda.api.requests.restaction.interactions.ReplyCallbackAction;
 import net.dv8tion.jda.internal.interactions.ModalInteractionImpl;
 
@@ -85,8 +85,8 @@ public class ModalSubmitInteractionEvent extends GenericInteractionCreateEvent i
 
     @Nonnull
     @Override
-    public List<TextInputMapping> getTextInputs()
+    public List<ModalMapping> getValues()
     {
-        return interaction.getTextInputs();
+        return interaction.getValues();
     }
 }
